:recycle: Move revealing votes to votes service and controller
