+++ issuepoker-backend/src/main/java/de/muenchen/issuepoker/issue/IssueController.java
@@ -4,7 +4,6 @@ import de.muenchen.issuepoker.entities.Issue;
 import de.muenchen.issuepoker.entities.dto.IssueDetailsDTO;
 import de.muenchen.issuepoker.entities.dto.IssueMapper;
 import de.muenchen.issuepoker.entities.dto.IssueRequestDTO;
-import de.muenchen.issuepoker.entities.dto.IssueRequestRevealedDTO;
 import de.muenchen.issuepoker.entities.dto.IssueSummaryDTO;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -47,11 +46,5 @@ public class IssueController {
         final Page<Issue> issuePage = issueService.getAllIssues(pageRequest);
         final List<IssueSummaryDTO> summaryList = issuePage.getContent().stream().map(issueMapper::toSummary).toList();
         return new PageImpl<>(summaryList, issuePage.getPageable(), issuePage.getTotalElements());
-    }
-
-    @PostMapping("{issueId}/revealed")
-    @ResponseStatus(HttpStatus.OK)
-    public void setRevealed(@PathVariable("issueId") final long issueId, @RequestBody final IssueRequestRevealedDTO revealedDTO) {
-        issueService.setRevealed(issueId, revealedDTO.revealed());
     }
 }
