+++ issuepoker-frontend/src/api/set-votes-revealed.ts
@@ -1,11 +1,11 @@
 import { defaultResponseHandler, postConfig } from "@/api/fetch-utils";
 
-export function setIssueRevealed(
+export function setVotesRevealed(
   issueId: string,
   revealed: boolean
 ): Promise<void> {
   return fetch(
-    `api/backend-service/issues/${issueId}/revealed`,
+    `api/backend-service/issues/${issueId}/votes/revealed`,
     postConfig({
       revealed: revealed,
     })
